Refactor items to their own route handlers
