+++ src/main/kotlin/HealthHandler.kt
@@ -1,0 +1,6 @@
+import io.javalin.http.Context
+object HealthHandler {
+    fun getHealth(ctx: Context) {
+        ctx.result("OK")
+    }
+}
