+++ src/main/kotlin/Server.kt
@@ -1,36 +1,29 @@
-import bikehopperclient.BikeHopperClient
-import bikehopperfilecreator.BikeHopperFileCreator
-import bikehopperfilecreator.BikeHopperFileCreatorException
 import io.javalin.Javalin
+import io.javalin.apibuilder.ApiBuilder.*
 import io.ktor.client.plugins.*
 import org.slf4j.LoggerFactory;
 
+import bikehopperfilecreator.BikeHopperFileCreatorException
 
 fun main() {
     val app = Javalin.create(/* TODO: Config */)
     val logger = LoggerFactory.getLogger("Server")
-    logger.info("Starting BikeHopperServer v0.1.3")
+    logger.info("Starting BikeHopperServer v0.1.4")
 
     app.before { ctx ->
         logger.info("Request: ${ctx.method()} ${ctx.path()} params: ${ctx.queryParamMap()}")
     }
 
-    app.get("/fit") { ctx ->
-        val bhClient = BikeHopperClient()
+    app.routes {
+        // Get a fit file
+        path("fit") {
+            get(FitFileHandler::getFitFile)
+        }
 
-        // Use bhClient to fetch our routes after pulling out the "path" param which is only used locally.
-        val routeData = bhClient.fetchRoute(ctx.queryParamMap().filter{ it.key != "path"})
-        // Quick check that the path we choose actually exists and has just a single leg.
-        // TODO: Figure out a better check eventually.
-        val path = ctx.queryParamAsClass<Int>("path", Int::class.java).check({it >= 0 && it < routeData.paths.size && routeData.paths[it].legs.size == 1 }, "Invalid path requested").get()
-        val bikeHopperFileCreator = BikeHopperFileCreator(routeData, path)
-        ctx.contentType("application/vnd.ant.fit")
-        ctx.result(bikeHopperFileCreator.getBuffer())
-    }
-
-    // Simple health check endpoint.
-    app.get("/health") { ctx ->
-        ctx.result("OK")
+        // Simple health check.  Returns OK on get.
+        path("health") {
+            get(HealthHandler::getHealth)
+        }
     }
 
     // Handle exceptions thrown from the BikeHopperClient's fetchRoute method
